Move external constraints to `validate` package
