+++ base/src/main/java/io/spine/validate/ExternalConstraints.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validate.constraint;
+package io.spine.validate;
 
 import com.google.common.base.Splitter;
 import com.google.common.collect.ImmutableSet;
